Remove axios workaround for another workaround so streaming works
Disabled Next.js's caching mechanism so that the response is actually streamed. Without this, the response would arrive in one big chunk in the stream reader code below. This behavior seems unique to Next.js. Tests in node and bun show that responses are streamed as expected regardless of the cache setting.